// src/app/interfaces/request/ClientRequest.ts
import { FormGroup } from "@angular/forms";

// Generated by https://quicktype.io
export interface ClientRequest {
    firstName:   string;
    lastName:    string;
    dni:         string;
    email:       string;
    city:        string;
    street:      string;
    number:      string;
    postalCode:  number;
    phoneNumber: string;
}

export function parseClientFormToRequest(formGroup: FormGroup): ClientRequest | null {
    if (!formGroup) {
        return null;
    }

    return {
        firstName: formGroup.get('firstName')?.value || null,
        lastName: formGroup.get('lastName')?.value || null,
        dni: formGroup.get('dni')?.value || null,
        email: formGroup.get('email')?.value || null,
        city: formGroup.get('city')?.value || null,
        street: formGroup.get('street')?.value || null,
        number: formGroup.get('number')?.value || null,
        postalCode: formGroup.get('postalCode')?.value || null,
        phoneNumber: formGroup.get('phoneNumber')?.value || null
    };
}