// src/app/interfaces/response/ClientResponse.ts
// Generated by https://quicktype.io
export interface SuccessClientResponse {
    statusCode: string;
    message:    string;
    url:        string;
    object:     ClientResponse;
}

export interface SuccessClientPageResponse {
    statusCode: string;
    message:    string;
    url:        string;
    object:     ClientPage;
}

export interface ClientPage {
    content:          ClientResponse[];
    pageable:         Pageable;
    last:             boolean;
    totalElements:    number;
    totalPages:       number;
    size:             number;
    number:           number;
    sort:             Sort;
    numberOfElements: number;
    first:            boolean;
    empty:            boolean;
}

export interface ClientResponse {
    id:          number;
    firstName:   string;
    lastName:    string;
    dni:         string;
    email:       string;
    city:        string;
    street:      string;
    number:      string;
    postalCode:  number;
    phoneNumber: string;
    membershipId:number;
    endingDate:  Date;
    status:      boolean;
}

export interface Pageable {
    pageNumber: number;
    pageSize:   number;
    sort:       Sort;
    offset:     number;
    paged:      boolean;
    unpaged:    boolean;
}

export interface Sort {
    empty:    boolean;
    sorted:   boolean;
    unsorted: boolean;
}